Make the sync tpu client use the async tpu client
